% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_moliRRBS.R
\name{format_moliRRBS}
\alias{format_moliRRBS}
\title{Format percent methylation from RRBS files for multi omic level integration moli}
\usage{
format_moliRRBS(
  x,
  meta,
  coverage = 0.8,
  window = 1000,
  values = c("beta", "M", "all")
)
}
\arguments{
\item{x}{A list of RRBS data frames}

\item{meta}{A metadata file with rownames matching names of x and a TRT column}

\item{coverage}{Coverage (enter as decimal, 0.8 = 80\%)}

\item{window}{Window of interest in bases upstream and downstream of genes}

\item{values}{Return beta (100*(M/(M+U)), gives beta distribution), M-values (log2((M+1)/(U+1)), Gaussian), or all}
}
\value{
A formatted .RDS file for use in moli, and csv of stats about features removed
}
\description{
Format percent methylation from RRBS files for multi omic level integration moli
}
