% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_RRBS.R
\name{format_RRBS}
\alias{format_RRBS}
\title{Format percent methylation from RRBS files for multi-omic integration}
\usage{
format_RRBS(x, meta, total_count = 10, coverage = 0.8, window = c(1000, 1000))
}
\arguments{
\item{x}{A list of RRBS data frames}

\item{meta}{A metadata file with rownames matching names of x and a TRT column}

\item{total_count}{Minimum counts (methylated + unmethylated) at each feature, default = 10}

\item{coverage}{Coverage (enter as decimal, 0.8 = 80\%)}

\item{window}{Window of interest in bases upstream and downstream of genes, requires upstream and downstream value, default = 1000, 1000}
}
\value{
Two formatted .RDS files of beta (100*(M/(M+U)) gives beta distribution) and M-values (log2((M+1)/(U+1)), Gaussian) for use in downstream analysis, and csv of stats about features removed.
}
\description{
Format percent methylation from RRBS files for multi-omic integration
}
